+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Unreleased
 - Adds `Overstrike.assign`.
+- Adds `Overstrike.isEmpty`.